[python] Split module.cpp
