+++ python/CMakeLists.txt
@@ -5,7 +5,7 @@ endif(GENERATE_PYTHON_STUBS)
 
 find_package(eigenpy 3.2.0 REQUIRED)
 
-set(pycandlewick_SOURCES src/module.cpp)
+set(pycandlewick_SOURCES src/expose-renderer.cpp src/module.cpp)
 
 set(PYLIB_INSTALL_DIR ${PYTHON_SITELIB}/${PROJECT_NAME})
